flask_api/requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
Flask-Marshmallow==1.2.2
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
Authlib==1.3.1
gunicorn==22.0.0